backend/requirements.txt
fastapi==0.109.2
uvicorn==0.30.5
pydantic==2.6.4
coverage==7.2.0
pylint==3.0.1
pytest==8.3.2
pytest-cov==5.0.0
pytest-asyncio==0.23.8
pyjwt==2.8.0